pyproject.toml
[project]
name = "ctfeed"
version = "0.1.0"
description = "Discord bot for tracking CTF events from CTFtime"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "py-cord>=2.4.1",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "asyncio-throttle>=1.0.2",
    "pytz>=2023.3",
    "audioop-lts>=0.2.2",
    "sqlalchemy>=2.0.45",
    "aiosqlite>=0.22.1",
    "pydantic-settings>=2.12.0",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "asyncpg>=0.31.0",
    "itsdangerous>=2.2.0",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "bandit>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
